Separated jqm-runner-api from hte api package and created its own package + did the refactoring that goes with this change
