+++ jqm-all/jqm-engine/src/main/java/com/enioka/jqm/engine/RunnerManager.java
@@ -5,7 +5,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import com.enioka.jqm.api.JobRunner;
+import com.enioka.jqm.runner.api.JobRunner;
 import com.enioka.jqm.jdbc.DbConn;
 import com.enioka.jqm.model.JobInstance;
 import com.enioka.jqm.runner.java.JavaRunner;
