+++ jqm-all/jqm-runner/jqm-runner-api/src/main/java/com/enioka/jqm/runner/api/JobRunner.java
@@ -1,5 +1,6 @@
-package com.enioka.jqm.api;
+package com.enioka.jqm.runner.api;
 
+import com.enioka.jqm.api.JobManager;
 import com.enioka.jqm.model.JobInstance;
 
 /**
