+++ jqm-all/jqm-runner/jqm-runner-api/src/main/java/com/enioka/jqm/runner/api/package-info.java
@@ -1,3 +1,3 @@
 @org.osgi.annotation.bundle.Export
 @org.osgi.annotation.versioning.Version("3.0.0")
-package com.enioka.jqm.api;
+package com.enioka.jqm.runner.api;
