+++ jqm-all/jqm-runner/jqm-runner-java/src/main/java/com/enioka/jqm/runner/java/JavaJobInstanceTracker.java
@@ -24,11 +24,11 @@ import java.util.Calendar;
 import javax.management.MBeanServer;
 import javax.management.ObjectName;
 
-import com.enioka.jqm.api.JobInstanceTracker;
+import com.enioka.jqm.runner.api.JobInstanceTracker;
 import com.enioka.jqm.api.JobManager;
-import com.enioka.jqm.api.JobRunnerCallback;
+import com.enioka.jqm.runner.api.JobRunnerCallback;
 import com.enioka.jqm.api.JobRunnerException;
-import com.enioka.jqm.api.JqmKillException;
+import com.enioka.jqm.runner.api.JqmKillException;
 import com.enioka.jqm.jdbc.DbConn;
 import com.enioka.jqm.model.Instruction;
 import com.enioka.jqm.model.JobInstance;
