+++ jqm-all/jqm-runner/jqm-runner-java/src/main/java/com/enioka/jqm/runner/java/JavaRunner.java
@@ -1,9 +1,9 @@
 package com.enioka.jqm.runner.java;
 
-import com.enioka.jqm.api.JobInstanceTracker;
+import com.enioka.jqm.runner.api.JobInstanceTracker;
 import com.enioka.jqm.api.JobManager;
-import com.enioka.jqm.api.JobRunner;
-import com.enioka.jqm.api.JobRunnerCallback;
+import com.enioka.jqm.runner.api.JobRunner;
+import com.enioka.jqm.runner.api.JobRunnerCallback;
 import com.enioka.jqm.jdbc.DbConn;
 import com.enioka.jqm.model.JobDef.PathType;
 import com.enioka.jqm.model.JobInstance;
